+++ test/Api.Test/AdminConsole/Authorization/Providers/Requirements/ManageProviderUsersRequirementTests.cs
@@ -1,18 +1,18 @@
-﻿using Bit.Api.AdminConsole.Authorization.Requirements;
+﻿using Bit.Api.AdminConsole.Authorization.Providers.Requirements;
 using Bit.Core.AdminConsole.Context;
 using Bit.Core.AdminConsole.Enums.Provider;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
 using Xunit;
 
-namespace Bit.Api.Test.AdminConsole.Authorization.Requirements;
+namespace Bit.Api.Test.AdminConsole.Authorization.Providers.Requirements;
 
 [SutProviderCustomize]
-public class ProviderUserRequirementTests
+public class ManageProviderUsersRequirementTests
 {
     [Theory, BitAutoData]
     public void Authorize_WhenUserIsProviderAdmin_ThenRequestShouldBeAuthorized(
-        SutProvider<ProviderUserRequirement> sutProvider)
+        SutProvider<ManageProviderUsersRequirement> sutProvider)
     {
         var providerClaims = new CurrentContextProvider { Id = Guid.NewGuid(), Type = ProviderUserType.ProviderAdmin };
 
@@ -22,19 +22,19 @@ public class ProviderUserRequirementTests
     }
 
     [Theory, BitAutoData]
-    public void Authorize_WhenUserIsServiceUser_ThenRequestShouldBeAuthorized(
-        SutProvider<ProviderUserRequirement> sutProvider)
+    public void Authorize_WhenUserIsServiceUser_ThenRequestShouldBeDenied(
+        SutProvider<ManageProviderUsersRequirement> sutProvider)
     {
         var providerClaims = new CurrentContextProvider { Id = Guid.NewGuid(), Type = ProviderUserType.ServiceUser };
 
         var actual = sutProvider.Sut.Authorize(providerClaims);
 
-        Assert.True(actual);
+        Assert.False(actual);
     }
 
     [Theory, BitAutoData]
     public void Authorize_WhenUserIsNotProviderMember_ThenRequestShouldBeDenied(
-        SutProvider<ProviderUserRequirement> sutProvider)
+        SutProvider<ManageProviderUsersRequirement> sutProvider)
     {
         var actual = sutProvider.Sut.Authorize(null);
